Fix not counting basilisk womb births towards times given birth
When giving birth to basilisks for Benoit(e) with the Basilisk Womb, it didn't count towards 'Times given birth'. Fixed that.

For the sake of DRY-coding I've moved the code to update the 'Birthed'-counter into a separate function called by 'player.updateBirthedCount();' and replaced the snippets with a call to it.