Fix poltergeist legs TF
Poltergeist legs TF from using Protoplasm was kinda borked, since it looped though LowerBody.HUMAN → LowerBody.GHOST → LowerBody.GHOST_2 over and over again, wasting a lot of Protoplasms.
Fixed that by making it so that Ectoplasm and Protoplasm doesn't change the lowerbody anymore, once it reached LowerBody.GHOST_2 (or LowerBody.GHOST for Ectoplasm).
I was considering about making Ectoplasm change LowerBody.GHOST_2 back to LowerBody.GHOST, but then decided, that this is not really needed and for those, who want to be an eldritch poltergeist with ghost-legs, there's always soulforce to fix that.